use c3i to submit execute pipelines, for unique names
