+++ tests/data/config-test/config.yml
@@ -4,7 +4,7 @@ aws-secret-key: your-secret
 aws-region-name: your-region
 concourse-url: your-server
 concourse-team: your-team
-concourse-user: your-user
+concourse-username: your-user
 concourse-password: your-user
 recipe-repo: your-repo
 # TODO: these need to be dynamically definable based on PR info
